+++ __tests__/event-spec.js
@@ -392,7 +392,7 @@ describe('Event', () => {
           const eventJSON = {
             id: 'id-1234',
             title: 'test event',
-            when: { time: 1409594400 },
+            when: { time: 1409594400, object: 'time' },
             participants: [
               {'name': 'foo', 'email': 'bar', 'status': 'noreply'}
             ],
@@ -407,6 +407,7 @@ describe('Event', () => {
           expect(event.id).toBe('id-1234');
           expect(event.title).toBe('test event');
           expect(event.when.time).toEqual(1409594400);
+          expect(event.when.object).toEqual('time');
           expect(event.iCalUID).toBe('id-5678');
           let participant = event.participants[0];
           expect(participant.toJSON()).toEqual(